refactor!: rework all our eslint configs to be clearer (#145)
* refactor!: rework all our eslint configs to be clearer

Do away with the confusing system of config overrides and dependencies.
Instead, define all 3 configs statically, which makes it much easier to
reason about what the linter is checking for and makes it much easier to
make changes. We've been at this for 5+ years, so we know what rules
we want for our repos, which means these external dependencies are more
of a hindrance than a help.

Going forward:
  * .js files will get the "arista-js" config
  * .jsx files will get the "arista-js" and "arista-react" configs
  * .ts files will get the "arista-ts" config
  * .tsx files will get the "arista-ts" and "arista-react" configs

The big difference here is that the "arista-js" config will no longer be
applied to TypeScript files. There are enough differences in the rules
that it warrants a separate config with no inheritance.

Move all common dev dependencies up to the root and remove all
dependencies (dev and prod) from all packages. Do away with anything
related to airbnb, flowtype, or prettier-config.

Remove jsx-a11y for now because this area needs an overhaul. We only had
one a11y rule enabled anyway, so we're not losing much here. There will
be an a11y audit this week so we may tackle this separately in the
future.

BREAKING CHANGE: all configs follow a new structure now (see above)

* Removed .eslintrc.json files in favor of the `-c` flag