fix(eslint-plugin-arista): rip out unneeded rules and buggy auto-fix
eslint 7.22.0 will throw errors if a plugin suggests an auto-fix that
has an invalid range, and that is now happening in both of our repos.

Due to long import names and the increasing use of TypeScript in our
code, it's now quite common to have comments in our import blocks, and
it appears that this plugin was never intended to handle them, thus
resulting in it suggesting impossible fixes. Seeing as how none of our
devs has ever complained about this, it seems that no one even uses the
auto-fix functionality.

Given this, and the fact that it would be annoying to fix, and the fact
that its utility is low, ripping out the buggy auto-fix code seems like
the best route. Developers can easily address these errors by hand.

On top of that, rip out rules that should never have been added in the
first place (like checking for multiple blank lines) because they
aren't related to import order and because they duplicate behavior
already found in prettier.